refactor: Move Enums to separate namespace.
