+++ src/Intranet.Selenium.Framework/Enums/AssertLevel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Intranet.Selenium.Framework
+namespace Intranet.Selenium.Framework.Enums
 {
     public enum AssertLevel
     {
