+++ src/Intranet.Selenium.Framework/ResultTracker.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Intranet.Selenium.Framework.Enums;
 
 namespace Intranet.Selenium.Framework
 {
