+++ src/Intranet.Selenium.Tests/FrameworkDiagnostics.cs
@@ -4,11 +4,12 @@ using System.Text;
 using Xunit;
 using OpenQA.Selenium;
 using Intranet.Selenium.Framework;
+using Intranet.Selenium.Framework.Enums;
 using System.Linq;
 
 namespace Intranet.Selenium.Tests
 {
-#pragma warning disable S3881 //full implementation of IDisposable not necessary; only used as a workaround (since xUnit does not support [TearDown] or similar)
+//pragma warning disable S3881 //full implementation of IDisposable not necessary; only used as a workaround (since xUnit does not support [TearDown] or similar)
     public class FrameworkDiagnostics : IDisposable
     {
         const string Url = "http://34.248.135.81";
@@ -77,10 +78,43 @@ namespace Intranet.Selenium.Tests
             driver.Log.Write("Test Completed");
         }
 
-        public void Dispose()
+        private bool disposedValue = false; // To detect redundant calls
+
+        protected virtual void Dispose(bool disposing)
         {
-            driver.Kill();
-            result.Evaluate();
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    // TODO: dispose managed state (managed objects).
+                }
+
+                // TODO: free unmanaged resources (unmanaged objects) and override a finalizer below.
+                // TODO: set large fields to null.
+
+                disposedValue = true;
+            }
         }
+
+        // TODO: override a finalizer only if Dispose(bool disposing) above has code to free unmanaged resources.
+        // ~FrameworkDiagnostics() {
+        //   // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+        //   Dispose(false);
+        // }
+
+        // This code added to correctly implement the disposable pattern.
+        void IDisposable.Dispose()
+        {
+            // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+            Dispose(true);
+            // TODO: uncomment the following line if the finalizer is overridden above.
+            // GC.SuppressFinalize(this);
+        }
+
+        //public void Dispose()
+        //{
+        //    driver.Kill();
+        //    result.Evaluate();
+        //}
     }
 }
